+++ Resources/public/js/components/products/components/edit/variants-list/main.js
@@ -287,8 +287,6 @@ define([
         };
 
     return {
-        view: true,
-
         templates: ['/admin/product/template/product/variants'],
 
         /**